+++ components/molecule/autosuggest/src/components/SingleSelection/index.js
@@ -4,7 +4,7 @@ import {Children} from 'react'
 import AtomInput from '@s-ui/react-atom-input'
 import MoleculeDropdownList from '@s-ui/react-molecule-dropdown-list'
 
-import {InputWithClearUI} from '../InputWithClearUI/index.js'
+import InputWithClearUI from '../InputWithClearUI/index.js'
 
 const MoleculeAutosuggestSingleSelection = ({
   autoFocus,