+++ components/molecule/imageEditor/src/index.js
@@ -1,12 +1,21 @@
 import {useCallback, useState} from 'react'
-import Cropper from 'react-easy-crop'
 
 import PropTypes from 'prop-types'
 
-import AtomSlider from '@s-ui/react-atom-slider'
+import {debounce} from '@s-ui/js/lib/function/debounce.js'
+import Injector from '@s-ui/react-primitive-injector'
 
 import getCroppedImg from './utils/cropImage.js'
-import {baseClass, DEFAULT_ASPECT, noop} from './config.js'
+import ImageEditorDefault from './ImageEditorDefault.js'
+import ImageEditorCropper from './ImageEditorCropper.js'
+import ImageEditorSliders from './ImageEditorSliders.js'
+import {
+  baseClass,
+  DEBOUNCING_TIME,
+  DEFAULT_ASPECT,
+  noop,
+  getRotationDegrees
+} from './config.js'
 
 const MoleculeImageEditor = ({
   aspect = DEFAULT_ASPECT,
@@ -16,82 +25,55 @@ const MoleculeImageEditor = ({
   onChange,
   onCropping = noop,
   rotateLabelIcon,
-  rotateLabelText
+  rotateLabelText,
+  children = <ImageEditorDefault />
 }) => {
-  const [crop, setCrop] = useState({x: 0, y: 0})
-  const [rotation, setRotation] = useState(0)
-  const [zoom, setZoom] = useState(0)
+  const [crop, cropSetter] = useState({x: 0, y: 0})
+  const [rotation, rotationSetter] = useState(0)
+  const [zoom, zoomSetter] = useState(0)
 
-  const getRotationDegrees = rotation => (rotation * 360) / 100
+  const setCrop = debounce(cropSetter, DEBOUNCING_TIME)
+  const setRotation = debounce(rotationSetter, DEBOUNCING_TIME)
+  const setZoom = debounce(zoomSetter, DEBOUNCING_TIME)
+
+  const cropCompleteHandler = async (croppedArea, croppedAreaPixels) => {
+    const rotationDegrees = getRotationDegrees(rotation)
+    onCropping(true)
+    const [croppedImageUrl, croppedImageBlobObject] = await getCroppedImg(
+      image,
+      croppedAreaPixels,
+      rotationDegrees
+    )
+    onChange(croppedImageUrl, croppedImageBlobObject)
+    onCropping(false)
+  }
 
   const onCropComplete = useCallback(
-    async (croppedArea, croppedAreaPixels) => {
-      const rotationDegrees = getRotationDegrees(rotation)
-      onCropping(true)
-      const croppedImage = await getCroppedImg(
-        image,
-        croppedAreaPixels,
-        rotationDegrees
-      )
-      onChange(croppedImage)
-      onCropping(false)
-    },
+    debounce(cropCompleteHandler, DEBOUNCING_TIME),
     [rotation, onCropping, image, onChange]
   )
 
   return (
     <div className={baseClass}>
-      <div className={`${baseClass}-crop`}>
-        <Cropper
-          image={image}
-          crop={crop}
-          zoom={1 + zoom / 100}
-          rotation={getRotationDegrees(rotation)}
-          aspect={aspect}
-          onCropChange={setCrop}
-          onCropComplete={onCropComplete}
-          onRotationChange={setRotation}
-          onZoomChange={zoom => setZoom((zoom - 1) * 100)}
-        />
-      </div>
-      <div className={`${baseClass}-slider`}>
-        {(cropLabelIcon || cropLabelText) && (
-          <div className={`${baseClass}-label`}>
-            {cropLabelIcon && (
-              <span className={`${baseClass}-labelIcon`}>{cropLabelIcon}</span>
-            )}
-            {cropLabelText && (
-              <span className={`${baseClass}-labelText`}>{cropLabelText}</span>
-            )}
-          </div>
-        )}
-        <AtomSlider
-          onChange={(event, {value}) => setZoom(value)}
-          value={zoom}
-          hideMarks
-        />
-      </div>
-      <div className={`${baseClass}-slider`}>
-        {(rotateLabelIcon || rotateLabelText) && (
-          <div className={`${baseClass}-label`}>
-            {rotateLabelIcon && (
-              <span className={`${baseClass}-labelIcon`}>
-                {rotateLabelIcon}
-              </span>
-            )}
-            {rotateLabelText && (
-              <span className={`${baseClass}-labelText`}>
-                {rotateLabelText}
-              </span>
-            )}
-          </div>
-        )}
-        <AtomSlider
-          onChange={(event, {value}) => setRotation(value)}
-          value={rotation}
-          hideMarks
-        />
-      </div>
+      <Injector
+        image={image}
+        crop={crop}
+        zoom={zoom}
+        rotation={rotation}
+        aspect={aspect}
+        onCropChange={setCrop}
+        onCropComplete={onCropComplete}
+        onRotationChange={setRotation}
+        onZoomChange={zoom => setZoom((zoom - 1) * 100)}
+        rotateLabelIcon={rotateLabelIcon}
+        rotateLabelText={rotateLabelText}
+        cropLabelIcon={cropLabelIcon}
+        cropLabelText={cropLabelText}
+        onZoomSliderChange={(event, {value}) => setZoom(value)}
+        onRotateSliderChange={(event, {value}) => setRotation(value)}
+      >
+        {children}
+      </Injector>
     </div>
   )
 }
@@ -105,7 +87,13 @@ MoleculeImageEditor.propTypes = {
   onChange: PropTypes.func.isRequired,
   onCropping: PropTypes.func,
   rotateLabelIcon: PropTypes.node,
-  rotateLabelText: PropTypes.string
+  rotateLabelText: PropTypes.string,
+  children: PropTypes.node
 }
 
 export default MoleculeImageEditor
+
+export {
+  ImageEditorCropper as MoleculeImageEditorCropper,
+  ImageEditorSliders as MoleculeImageEditorSliders
+}